llm_doc_processor/requirements.txt
fastapi
uvicorn
spacy
python-dotenv

faiss-cpu
pdfplumber
python-docx
sentence-transformers

python-dotenv
google-generativeai
